+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- `mirror_defaults` decorator for mirroring the default arguments of another
+  function.
+
 ## [0.3.0] - 2019-06-10
 
 ### Added